Fix convergence error counting in snapshot simulations
If a snapshot simulation was configured with
"Convergence error percent threshold" = 1.0 and the single time point
resulted in a convergence error then the simulation did not fail with
that error.

A 1.0% error threshold with one time point means that 0 errors are
allowed. The bug was that the code was using 0 to indicate that the
user did not want to define a threshold. Fixed the bug by using None to
indicate the case of no threshold.