Fix registry management
This fixes a problem that occurs when a user uses PyDSS, implicitly
serializing default controllers, and then installs PyDSS in a new
location, such as with a new version of Python. PyDSS would try to load
controllers from an invalid path.

This fixes the issue by never serializing default controllers.