Fix data export bug when devices are disabled
The OpenDSS element iteration process in the data export code was not
correctly handling the case when devices were disabled in the middle of
a simulation. OpenDSS does not return disabled devices and so the code
was exporting the values for the last enabled device multiple times.
That could have resulted in invalid data (worst case) or crashes
(best case) if the dimensional size of the data was different across
enabled and disabled elements.