Move aiohttp package to extras installation
Inclusion of this package in the default installation sometimes causes
Windows users to experience failures. aiohttp requires the package
multidict, which at least sometimes has to built from source. If the
Windows user does not already have the correct version of Microsoft C++
Build Tools installed, the process will fail. The user then has to
install those tools, which involves downloading a 9 GB package. This is
onerous and not usually necessary because most PyDSS users do not need
the aiohttp package.